Smarty assign missing 'templatePath' globally
The next step is to check this mainly in every theme that has a preview_iframe.tpl file itself. But this might also play a role in the frontend.

Remember, frontends preview_iframe.tpl on save is called, when backends template 2k11 is defined.
In this case the template is already used for the parent and why {serendipity_getFile ...} will not work, since searching in '2k11' instead of the {$template} theme.
You could workaround this by using {serendipity_getFile file="../templateName/asses"}, which would bubble-up and then go to the user template.
Therefore it is better to use {$serendipityHTTPPath}{$templatePath}{$template} instead, since you really want template only assets.
(Well, certainly this only, when you need them for the preview file or in your frontend. You should always check this first, to not load and request anything you don't need in mode 'save' or 'preview'.)