plugin manager: Workaround upgradable plugins db caching
for the remote changelog path per SESSION storage.

By time first, the plugin manager is reading the xml file, which gets the correct foreignPlugin changelog path, added by the 'Automagic XML sync', documented here: 93da2eb.
If not performing the update immediately this data gets lost, since it cannot be stored adequate to the current pluginlist table fields.
As long as not closing the browser, this remote ChangeLog path is now available per SESSION storage.

(Avoiding Cookie storage for this matter to survive destroying the current session... though we still might have to move to, if the session approach is not applicable.)

Todo: Remove that SESSION data if the upgrade went through in time.