ml: Set media gallery list item head title to diskname for local saved files
Inserting doublete named images, by reloading accident or explicit commission, causes the media gallery list item head title to show up with the origin file name, which is held and kept in the 'realname' database field.
Since the storage name, disk-file-name and the thumb get incremented names like imagename1.jpg imagename2.jpg etc for doubletes, this is irritating for not so deeply involved users. See L326 $realname = serendipity_imageAppend($tfile, $target, .....) and ff.

So we switch diskname and realname list appearance, since we want to show the real existing file name, which actually is the saved diskname. The realname, as the book states, is for the origin name.