spamblock: Move link into sidebar section SETTING, since being one
Though, its a question of definition. It either is an activity, eg. to emergency stop spam floods, or a special highlighted plugin setting for a general backend behaviour.
Since it was previously added next to comments it was seen more like a comment bundle action, but the comments page has its own button for spamblock quickjumps.