Improve ac5a357 removing USERLEVEL and adding a NEW siteAutoUpgraders
permission flag for multiuser systems.

This allows to give a "special" extending ADMINISTRATOR right to an
"Editor in CHIEF" labeled user, without borking the CHIEFs default
permisssions.
The best and recommended approch is to build a special group with this
flag enabled and assign it to the Chief user you want it to have.
To not allow other Chiefs or Users to access this group to other users,
you have to additionally set it as a 'hiddenGroup / Non-user" inside the
new created group.

This will start working with your next Styx update and a next update of the autoupdate plugin, as v.1.5.8.