templates: Allow backend themes "gallery" using WebP formated images
If a WebP image isn't available in the theme or the browser used does not support it,
it has a fallback to the origin preview.png and/or preview_fullsize.jpg files.
The fallback even distinguishes for the case that the preview_fullsize.jpg is less in
size than the smaller preview.png file, which is possible with the png file format.

Browser support, see https://caniuse.com/webp