ml: Fix s*_scaleImg() return checks for 80512f3 26d41b6 779adbe 4f93556
This is part one of a general rework of return checks, since we have to many confusing back handles.
This is for: serendipity_resizeImageGD(), serendipity_passToCMD() and
serendipity_convertToWebPFormat() returns.
They shall all give back an array and the [0] key as 0, which is success!
Else a $result[0] != 0 is an error or failed execution!