ml: Fix return checks part II see 582b386
This is part II of a general rework of return checks, since we had to many confusing back handles.
This is for: serendipity_resizeImageGD(), serendipity_passToCMD() and
serendipity_convertToWebPFormat() returns.
They shall all give back an array and the [0] key as 0, which is success!
Else a $result[0] != 0 is an error or failed execution!

Needs a deep check and test run to see if I fetched them all...!