ml: Fixed rename issue for the full webp variation file
in case it was uploaded as "example.png.jpg" and afterwards renamed to "example.jpg".
In this case the pathinfo part removed the necessary first ".png" extension in the file name, while being an extension and not recognizing it being a filename part.
Disabling this caused a regression with the hidden directory paranoid check,
which now is more precise about the string to check for.
Well done!
I think alpha2 may be ready now! :)