entries API: Fix stable Archive assigning wrong pagination values
so any themes entry paginations got wrong start values for the current page,
and the page calculations for prev/next pages.

Rant:
This kind of fucked-up code - never responsibly checked and cared in deep -
drove me away from Origin to build a personal cared Serendipity of what it should be!