compat: Partially sync a (garvin) (s9y) comment for ccfadb7 ff
and try to work out where I find differences as guidelines!

Remember, the origin issue which caused the 1729feb workaround, was somebody
willingly forcing a serendipity GET variable NOT as an array, which just caused
some PHP warnings on later array handlers, which for itself isn't bad
and did not seem to open a vulnerable on it.