images: Follow camelCase principle
Function names are case-insensitive, so actually there is no need to change them immediately in imageslectorplus, responsiveimages, usergallery plugins.
Still not absolutely sure about possible (internal) conflict with Smarty registered
$serendipity['smarty']->registerPlugin('function', 'serendipity_getImageSize', 'serendipity_smarty_getImageSize');
This wasn't a problem before, so this probably is capsulated and/or namespaced, or such... and is out of way.