routing: Remove 3079caf hotfix
which was added for freetag autocomplete scriptlets in combo with staticpage plugin usage, which then threw a 404.
This is not the case any more AFAICS!
Additionally calling genpage is now limited to frontend calls only.
Also adds 'empty' to the GET action call which is a valid setting to not push/run default actions on this. Thanks to Stephan Brunker to find out.