default: Improve preset styles part two
Removes too unstable parts in f0836cf ff, which were not that helpful.
Regard; Media Galleries by columns need more items than the selected columns.
Testing with c3 columns and 3 items will do in firefox, but not in chromium without expanding the box for possible more.
And this needs to be display: inline-box, not flexed, else every item is flexed into one line.