plugins: Fix Spartacus Plugin Synchronisation - Part IV -
sidebar plugin upgrades.

See part I-III commits.

Now that the wrong locationtype was fixed, the Spartacus Plugin Synchronisation
could do its job, purging and creating the pluginlist items correctly,
but now revealed missing another setPluginInfo() run for the local sidebar
list to checkout the 'upgrade_version' for being 'upgradeable'.
We better just use the direct and safe way to place this item to the
database again.
This runs on upgrades and/or opening the sidebar install list for the very
most plugins once only.