history: Simplify and improve looped years caching checks
Due filemtime($cachefile) written by fopen() mode 'w' would need a
following touch($cachefile) for a changing modification date (which
is filemtime), when not having been unlinked before. But this is not
really a need at all, since the 'w' mode sets the content to zero and
creates a new file if required.