b46: Allow featured article preview on entries paging archives
Remove
in_array($view, ['start', 'entries', 'archives'])
 to
in_array($view, ['start', 'entries'])
if you don't want this behaviour.

It is implemented to allow having it on /P1.html, as this should be the same as the 'start' $view, but - while paging - is of $view type 'archives'.
To exclude P2+ views if would need additional logic, making it complicated.