templates: Experimental move 2k11 core theme a tad down in sorting
Best would be to rename it to something natural, since PHPs sort algorithm
is putting starting numerics either to top, which is somehow like using
!important in CSS, or to bottom sorts by flags.
But I still like the name and it has been long time the flagship standard,
so we should keep that honoured.
By this terms I place it left to Next.