plugin_config_item: Improve input field containers placement and by div
for active/open has_info additional_info containers, to not "jump" by margin.
Example was the good media_chooser / radio items, which showed to besser use a div.

Still open issues are textarea and selection container items with a has_info
container. We try to solve this by CSS next.