smarty: Temporarily fix Styx 'debug' with Smarty debugging = true case
where both template data array keys are not set (initialized).

This is caused by our Styx smarty class parameter "debugging = true;"
This failure was checked against a templates {debug} which does not have this issue.
It is not clear which rendered template call is the root cause, since we have differing call use cases and it does not do this everywhere.
I found it appear only in entry and staticpage entry forms. Maybe that gives a hint which specific call causes this.