boot: Remove standard contactform templates $required_fields array
variable usage. This array is meant for real comment form fields only, and spiced by the spamblock and commentspice plugins.
Sadly it has been added by copy & paste to nearly all these standard contactform templates in themes and makes it hard for a
normal user to understand why it is only working on some contact form fields, because of being "evolutionarily" related to comment fields and handlers.
The default required fields for the standard contactform template are always: name, email, comment (message)!
Else this is configurable by using the other form presets and the "Form field string" with adding 'require' or not.