images API: Add temporary HOTFIX for missing getimagesize() AVIF size
An ugly workaround using the existing WebP file expression for the sizes,
since currently PHP 8.1 getimagesize() won't work without access
to libgd or libavif as external libs.
Google DEVs are currently working on a build-in function,
being able to detect image size for AVIF files. If they succeed
this will sooner or later be merged.