images API: Remove _displayImageList() obsolete $lineBreak parameter
as from Series 0/1 (?) for an issue with the ML sort filter [perpage] set to 9, which gave 10 because of used Modulo.

For now we keep the depending $lineBreak parameter for _showMedia() but that approach looks like being obsolete too and probably will be purged soon in a second step, since I can't seem to find a reason of some external file/plugins or other rare cases using these variables (any more).