smarty: [Sync 4.2.0] Templatelexer in the main points
Most of it was borked an thus not merged (and needs a cleanup in Smarty origin).

This Styx revision keeps "$this->yy_global_pattern1" pattern structure
for the lexer tokens to work with our XML/PHP templates. (Leaving it like this until I really get was this is doing!)
We just simply "deactivate" unsupported php code patterns for the marty dropped PHP support.