importer::wordpress-pg: Trying to follow user fixes of 9d6faf0c0106126d83aca8809ef7cb90745bc2f3
 - Fix connection string variables

Coding in the shadow of not having a PostgreSQL test environment!!!!!

This WP-PG seems to miss some of the "features" that the WP (mysqli) variant has. So this needs a closer look!

Is there a mysqli_report(MYSQLI_REPORT_OFF) equivalent (no exception failure) for PostgreSQL when asking for tables that don't exist ?