entries API: On search following 6b7b956a3f16f375ced47ef1a40420047d990df4 with PostgreSQL
there are other issue when to_tsvector is not available and the ILIKE fallback is used.
I tried to fix it this way since the wildcard search retry can't be used and the checkcount is actually a MySQL thing (limitation) only.
This may not be the last word of wisdom, but it works for now.