entries API: Improve and fix the 'frontend_fetcharchives' "entries per
month" counters SQL query in s*_printArchives() for the categorytemplates plugin.

This in special is the case for this categorytemplates plugin only, which JOINs the entrycat table to the relative simple entries counter SQL for to exclude any hidden set categorytemplates.
This JOIN makes it mandatory to have a DISTINCT (GROUP BY) distinction to support entries being assigned to multi categories NOT being counted multiple times.