spartacus: Prepare having switched additional_plugins/ to legacy branch
This will allow us to support security and important bugfix commits to the Series 4 a little longer until Styx 5.0 is stable enough and widely used.
Styx 5 Series will point to a dev/ branch or even again use the master/ branch for development time, which isn't decided yet.

Spartacus website will be on legacy that long.