smarty API: Fix 8ee72a52ceec8d8d54db13a5c5c6efc5916dbc98 since registering PHP methods like this (natively) is not working
This might be an issue in the Smarty checkup to throw deprecation notices
for upcoming Smarty 5, which popup once to notify that these are deprecated,
but since we need this once only for the freetag plugin hook into entries.tpl,
we can get around it using it at all. All themes checked.