lang API: Fixup dd832254472a702921fdb9f7bbbe0ac4af687d2c#diff-9ab5e4fd4c9b654a7843bf0b989f8c4653595e31aa0875417172955a6a215464 for the undocumented private variable usage to overwrite lang constants via user theme
I think backend allowance was just an overseen mistake while this commit was a draft commit for splitting up backend and frontend themes for Serendipity 2.0.
It makes no sense to have this anywhere other than for individual user frontend themes. The less it runs through, the less possible regressions it has.
It was done in the early days to allow overwriting certain core constants in themes for frontend templates, e.g. POWERED_BY.