entries_admin API: Change default textarea names for id-naming-convention
to either be of sort camelCase, snake_case, PascalCase or kebab-case.
The default is snake_case, so to be 'serendipity_textarea_body', etc.
Using this array like 'serendipity[body]' naming may effort escaping.
This old style works well within attribute name cases, but not for the the id name.
And snake_case makes life easier! :-)