pure: Review of blockquote styles
The "flow-root" helps us to not break the spawned left flow in floated situations, like left floated image with right text containing a blockquote, else the blockquote orients itself at the "parents" content container left side and "breaks apart" when the floated image ends. Now it sticks together as long the blockquote is used, even when the images has reached its bottom. Same behavior as we implemented for lists.