config API: Fix serendipity_load_configuration() $author array offset for PHP 8.5
Deprecation up from PHP 8.5 RC2/RC3 - Using null as an array offset is deprecated, use an empty string instead.
Kepping the default value null is a matter of using ?int, so we can't use an empty string here.
Currently it seems more clean to me to turn null to an empty string for the offset once we are in.