config API: Improve a regression of 041acae61176c6c408a30645c4787dd885f51424 partial change for authorid fallback on null
Keep the null as before, which then throws a deprecation notice for unlogged-in user access in the frontend with some plugins and check it again on null right after to return false.
We do not need to go any further with a fake authorid = 0 to make the static keeping work if that has no real purpose to keep.